--- core-service/src/main/java/ru/ssemenov/controllers/StatisticsController.java
@@ -1,9 +1,0 @@
-package ru.ssemenov.controllers;
-
-import org.springframework.stereotype.Controller;
-import ru.ssemenov.dtos.StatisticsDto;
-
-@Controller
-public interface StatisticsController {
-    StatisticsDto getStatistics();
-}
