--- core-service/src/main/java/ru/ssemenov/services/StatisticsService.java
@@ -1,17 +1,0 @@
-package ru.ssemenov.services;
-
-import org.springframework.stereotype.Service;
-
-import java.time.LocalTime;
-
-@Service
-public interface StatisticsService {
-
-    LocalTime getAverageDurationOfRelease();
-
-    Double getShareOfSubmittedDeclarationsBeforeNoonByMonth();
-
-    Double getShareOfReleasedDeclarationsWithinOneDayByMonth();
-
-    Long getDeclarationsCountInProgress();
-}
